+++ match/match.sln
@@ -15,6 +15,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Piipan.Match.Orchestrator.T
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Piipan.Match.State.Tests", "tests\Piipan.Match.State.Tests\Piipan.Match.State.Tests.csproj", "{28F4A3A3-DC11-40F6-85C2-F6F5E9856004}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Piipan.Match.Shared", "src\Piipan.Match.Shared\Piipan.Match.Shared.csproj", "{EE488CE4-2137-4B75-9B46-2E16F2F5A41B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -76,11 +78,24 @@ Global
 		{28F4A3A3-DC11-40F6-85C2-F6F5E9856004}.Release|x64.Build.0 = Release|Any CPU
 		{28F4A3A3-DC11-40F6-85C2-F6F5E9856004}.Release|x86.ActiveCfg = Release|Any CPU
 		{28F4A3A3-DC11-40F6-85C2-F6F5E9856004}.Release|x86.Build.0 = Release|Any CPU
+		{EE488CE4-2137-4B75-9B46-2E16F2F5A41B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EE488CE4-2137-4B75-9B46-2E16F2F5A41B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EE488CE4-2137-4B75-9B46-2E16F2F5A41B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{EE488CE4-2137-4B75-9B46-2E16F2F5A41B}.Debug|x64.Build.0 = Debug|Any CPU
+		{EE488CE4-2137-4B75-9B46-2E16F2F5A41B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{EE488CE4-2137-4B75-9B46-2E16F2F5A41B}.Debug|x86.Build.0 = Debug|Any CPU
+		{EE488CE4-2137-4B75-9B46-2E16F2F5A41B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EE488CE4-2137-4B75-9B46-2E16F2F5A41B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EE488CE4-2137-4B75-9B46-2E16F2F5A41B}.Release|x64.ActiveCfg = Release|Any CPU
+		{EE488CE4-2137-4B75-9B46-2E16F2F5A41B}.Release|x64.Build.0 = Release|Any CPU
+		{EE488CE4-2137-4B75-9B46-2E16F2F5A41B}.Release|x86.ActiveCfg = Release|Any CPU
+		{EE488CE4-2137-4B75-9B46-2E16F2F5A41B}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{602212DC-6449-4E6C-A221-7464DC2B0909} = {CE0D7711-0AC2-48A0-BAD3-D826C058A84F}
 		{98DAC477-3C0F-45E1-BFB7-084F1BD0FF36} = {CE0D7711-0AC2-48A0-BAD3-D826C058A84F}
 		{065C892E-DF05-46BD-B859-FF457D488098} = {B28ACB9E-5540-4C32-AD0C-073B5579441E}
 		{28F4A3A3-DC11-40F6-85C2-F6F5E9856004} = {B28ACB9E-5540-4C32-AD0C-073B5579441E}
+		{EE488CE4-2137-4B75-9B46-2E16F2F5A41B} = {CE0D7711-0AC2-48A0-BAD3-D826C058A84F}
 	EndGlobalSection
 EndGlobal
